pyproject.toml
[tool.poetry]
name = "teg-technologie-generatywne"
version = "0.1.0"
description = ""
authors = ["Tomasz Serafiński <work.serafinski@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
langchain-openai = "^0.3.8"
langchain = "^0.3.20"
langchain-core = "^0.3.43"
notebook = "^7.3.2"
flask = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
